+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/transformers/inbound/providers/DecoratedSamlResponseToIdaResponseIssuedByIdpTransformer.java
@@ -4,12 +4,12 @@ import org.opensaml.saml.saml2.core.Assertion;
 import org.opensaml.saml.saml2.core.Response;
 import org.opensaml.saml.saml2.metadata.IDPSSODescriptor;
 import uk.gov.ida.saml.hub.domain.InboundResponseFromIdp;
+import uk.gov.ida.saml.hub.validators.response.idp.EncryptedResponseFromIdpValidator;
 import uk.gov.ida.saml.security.AssertionDecrypter;
 import uk.gov.ida.saml.security.SamlAssertionsSignatureValidator;
 import uk.gov.ida.saml.hub.transformers.inbound.IdaResponseFromIdpUnmarshaller;
 import uk.gov.ida.saml.hub.transformers.inbound.decorators.ResponseDestinationValidator;
-import uk.gov.ida.saml.hub.validators.response.ResponseAssertionsFromIdpValidator;
-import uk.gov.ida.saml.hub.validators.response.ResponseFromIdpValidator;
+import uk.gov.ida.saml.hub.validators.response.idp.ResponseAssertionsFromIdpValidator;
 import uk.gov.ida.saml.security.validators.ValidatedAssertions;
 import uk.gov.ida.saml.security.validators.ValidatedResponse;
 import uk.gov.ida.saml.security.validators.signature.SamlResponseSignatureValidator;
@@ -23,7 +23,7 @@ public class DecoratedSamlResponseToIdaResponseIssuedByIdpTransformer implements
     private final SamlResponseSignatureValidator samlResponseSignatureValidator;
     private final AssertionDecrypter assertionDecrypter;
     private final SamlAssertionsSignatureValidator samlAssertionsSignatureValidator;
-    private final ResponseFromIdpValidator responseFromIdpValidator;
+    private final EncryptedResponseFromIdpValidator responseFromIdpValidator;
     private final ResponseDestinationValidator responseDestinationValidator;
     private final ResponseAssertionsFromIdpValidator responseAssertionsFromIdpValidator;
 
@@ -32,7 +32,7 @@ public class DecoratedSamlResponseToIdaResponseIssuedByIdpTransformer implements
             SamlResponseSignatureValidator samlResponseSignatureValidator,
             AssertionDecrypter assertionDecrypter,
             SamlAssertionsSignatureValidator samlAssertionsSignatureValidator,
-            ResponseFromIdpValidator responseFromIdpValidator,
+            EncryptedResponseFromIdpValidator responseFromIdpValidator,
             ResponseDestinationValidator responseDestinationValidator,
             ResponseAssertionsFromIdpValidator responseAssertionsFromIdpValidator) {
