+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/validators/response/idp/EncryptedResponseFromIdpValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.saml.hub.validators.response;
+package uk.gov.ida.saml.hub.validators.response.idp;
 
 import com.google.common.base.Strings;
 import org.opensaml.saml.saml2.core.Issuer;
@@ -9,9 +9,9 @@ import org.opensaml.saml.saml2.core.StatusCode;
 import org.opensaml.xmlsec.signature.Signature;
 import uk.gov.ida.saml.core.validation.SamlTransformationErrorException;
 import uk.gov.ida.saml.core.validation.SamlValidationSpecificationFailure;
+import uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory;
 import uk.gov.ida.saml.hub.domain.IdpIdaStatus;
 import uk.gov.ida.saml.hub.transformers.inbound.SamlStatusToIdaStatusCodeMapper;
-import uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory;
 import uk.gov.ida.saml.hub.transformers.inbound.SamlStatusToIdpIdaStatusMappingsFactory;
 import uk.gov.ida.saml.security.validators.signature.SamlSignatureUtil;
 
@@ -20,26 +20,45 @@ import java.util.Optional;
 import static uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory.invalidStatusCode;
 import static uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory.invalidSubStatusCode;
 
-public abstract class ResponseFromIdpValidator {
+public class EncryptedResponseFromIdpValidator {
+
+    public EncryptedResponseFromIdpValidator(final SamlStatusToIdpIdaStatusMappingsFactory statusMappingsFactory) {
+        this.statusCodeMapper = new SamlStatusToIdaStatusCodeMapper(
+            statusMappingsFactory.getSamlToIdpIdaStatusMappings()
+        );
+    }
+
+    protected void validateAssertionPresence(Response response) {
+        if (!response.getAssertions().isEmpty()) {
+            SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.unencryptedAssertion();
+            throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
+        }
+
+        boolean responseWasSuccessful = response.getStatus().getStatusCode().getValue().equals(StatusCode.SUCCESS);
+        if (responseWasSuccessful && response.getEncryptedAssertions().isEmpty()) {
+            SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.missingSuccessUnEncryptedAssertions();
+            throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
+        }
+
+        if (!responseWasSuccessful && !response.getEncryptedAssertions().isEmpty()) {
+            SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.nonSuccessHasUnEncryptedAssertions();
+            throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
+        }
+
+        if (responseWasSuccessful && response.getEncryptedAssertions().size() != 2) {
+            SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.unexpectedNumberOfAssertions(2, response.getEncryptedAssertions().size());
+            throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
+        }
+    }
 
     private static final int SUB_STATUS_CODE_LIMIT = 1;
 
     private SamlStatusToIdaStatusCodeMapper statusCodeMapper;
 
-    protected ResponseFromIdpValidator(
-            final SamlStatusToIdpIdaStatusMappingsFactory statusMappingsFactory) {
-
-        this.statusCodeMapper = new SamlStatusToIdaStatusCodeMapper(
-                statusMappingsFactory.getSamlToIdpIdaStatusMappings()
-        );
-    }
-
     public void validate(Response response) {
         validateRequestIdAndIssuerId(response);
         validateResponse(response);
     }
-
-    protected abstract void validateAssertionPresence(Response response);
 
     private void validateRequestIdAndIssuerId(Response response) {
         Issuer issuer = response.getIssuer();
