+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/validators/response/matchingservice/HealthCheckResponseFromMatchingServiceValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.saml.hub.validators.response;
+package uk.gov.ida.saml.hub.validators.response.matchingservice;
 
 import com.google.common.base.Strings;
 import org.opensaml.saml.saml2.core.Issuer;
@@ -7,12 +7,11 @@ import org.opensaml.saml.saml2.core.Response;
 import org.opensaml.saml.saml2.core.StatusCode;
 import org.opensaml.xmlsec.signature.Signature;
 import uk.gov.ida.saml.core.domain.SamlStatusCode;
-import uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory;
 import uk.gov.ida.saml.core.validation.SamlTransformationErrorException;
 import uk.gov.ida.saml.core.validation.SamlValidationSpecificationFailure;
-import uk.gov.ida.saml.security.validators.signature.SamlSignatureUtil;
+import uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory;import uk.gov.ida.saml.security.validators.signature.SamlSignatureUtil;
 
-public abstract class ResponseFromMatchingServiceValidator {
+public class HealthCheckResponseFromMatchingServiceValidator {
 
     public void validate(Response response) {
         validateAndExtractRequestIdAndIssuerId(response);
@@ -59,10 +58,7 @@ public abstract class ResponseFromMatchingServiceValidator {
         }
         validateIssuer(response.getIssuer());
         validateStatusAndSubStatus(response);
-        validateAssertionPresence(response);
     }
-
-    protected abstract void validateAssertionPresence(Response response);
 
     protected void validateStatusAndSubStatus(Response response) {
 
@@ -79,25 +75,20 @@ public abstract class ResponseFromMatchingServiceValidator {
             throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
         }
 
-        boolean statusWasResponder = response.getStatus().getStatusCode().getValue().equals(StatusCode.RESPONDER);
-        boolean statusWasSuccess = response.getStatus().getStatusCode().getValue().equals(StatusCode.SUCCESS);
+        final String statusCodeValue = statusCode.getValue();
+        boolean statusWasSuccess = statusCodeValue.equals(StatusCode.SUCCESS);
 
-        boolean subStatusWasNoMatch = response.getStatus().getStatusCode().getStatusCode().getValue().equals(SamlStatusCode.NO_MATCH);
-        boolean subStatusWasMatch = response.getStatus().getStatusCode().getStatusCode().getValue().equals(SamlStatusCode.MATCH);
-        boolean subStatusWasMultiMatch = response.getStatus().getStatusCode().getStatusCode().getValue().equals(SamlStatusCode.MULTI_MATCH);
-        boolean subStatusWasCreated = response.getStatus().getStatusCode().getStatusCode().getValue().equals(SamlStatusCode.CREATED);
-        boolean subStatusWasCreateFailure = response.getStatus().getStatusCode().getStatusCode().getValue().equals(SamlStatusCode.CREATE_FAILURE);
+        final String subStatusCodeValue = statusCode.getStatusCode().getValue();
+        boolean subStatusWasHealthy = subStatusCodeValue.equals(SamlStatusCode.HEALTHY);
 
-        if (statusWasResponder) {
-            if (!subStatusWasNoMatch && !subStatusWasMultiMatch && !subStatusWasCreateFailure) {
-                SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.subStatusMustBeOneOf("Responder", "No Match", "Multi Match", "Create Failure");
-                throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
-            }
-        } else {
-            if (statusWasSuccess && !(subStatusWasMatch || subStatusWasNoMatch || subStatusWasCreated)) {
-                SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.subStatusMustBeOneOf("Success", "Match", "No Match", "Created");
-                throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
-            }
+        if (!statusWasSuccess) {
+            SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.invalidStatusCode(statusCodeValue);
+            throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
+        }
+
+        if (!subStatusWasHealthy) {
+            SamlValidationSpecificationFailure failure = SamlTransformationErrorFactory.invalidSubStatusCode(subStatusCodeValue, StatusCode.SUCCESS);
+            throw new SamlTransformationErrorException(failure.getErrorMessage(), failure.getLogLevel());
         }
     }
