+++ hub-saml/src/test/java/uk/gov/ida/saml/hub/validators/response/matchingservice/EncryptedResponseFromMatchingServiceValidatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.saml.hub.validators.response;
+package uk.gov.ida.saml.hub.validators.response.matchingservice;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -8,10 +8,11 @@ import org.opensaml.saml.saml2.core.Response;
 import org.opensaml.saml.saml2.core.Status;
 import org.opensaml.saml.saml2.core.StatusCode;
 import uk.gov.ida.saml.core.domain.SamlStatusCode;
+import uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory;
 import uk.gov.ida.saml.core.test.OpenSAMLMockitoRunner;
 import uk.gov.ida.saml.core.test.SamlTransformationErrorManagerTestHelper;
 import uk.gov.ida.saml.core.validation.SamlValidationSpecificationFailure;
-import uk.gov.ida.saml.core.errors.SamlTransformationErrorFactory;
+
 import static uk.gov.ida.saml.core.test.builders.AssertionBuilder.anAssertion;
 import static uk.gov.ida.saml.core.test.builders.IssuerBuilder.anIssuer;
 import static uk.gov.ida.saml.core.test.builders.ResponseBuilder.aResponse;
@@ -23,7 +24,7 @@ public class EncryptedResponseFromMatchingServiceValidatorTest {
 
     private Status happyStatus;
 
-    private ResponseFromMatchingServiceValidator validator;
+    private EncryptedResponseFromMatchingServiceValidator validator;
 
     @Before
     public void setUp() throws Exception {
